bqscrape: Implement GetAllTables the simple way
Gets the metadata for all tables in a dataset by fetching each table one at
a time. This will need to be parallelized to make it faster, but for this
will work for a first pass.

bqscrape_test.go: Simplify the fake backend
integrationtests: Add a very dumb test that connects to a real test project.